refactor: Modal 스토리 props별 분리
